select field refactored
